change method signiture - int to double

+++ src/main/java/domain/amount/InvestmentAmount.java
@@ -5,5 +5,5 @@ import domain.interest_rate.InterestRate;
 public interface InvestmentAmount {
 	double calAnnualInterest(InterestRate interestRate);
 
-	int calMonthlyInterest(InterestRate interestRate);
+	double calMonthlyInterest(InterestRate interestRate);
 }
